Changes versions os browsers to be tested in travis
Safari 8 is almost not used anymore, we should test on Safari 9 and 10 instead.

Saucelabs started throwing errors saying that the version of Microsoft Edge we're asking for isn't supported. That's because we were using the browsers version itself, instead of EdgeHTML's version (which is on version 14 now, hasn't reached 20 yet), which is the one they're expecting now.